+++ CHANGELOG.rst
@@ -6,6 +6,9 @@ and this project adheres to `Semantic Versioning <https://semver.org/spec/v2.0.0
 
 Latest
 ------
+
+0.11.0 (2025-02-12)
+------------------
 * Added Checklist and Observation Managers for convenience queries.
 
 0.10.0 (2025-02-11)